Refactor: Extract AudioManager and improve code organization
- Extract AudioManager (260 lines) from AudioPriorityBarApp.swift to ViewModels/
- Move StoredDevice struct to Models/StoredDevice.swift
- Convert isActiveOutputMuted/isActiveInputMuted to computed properties
- Add MARK comments for better code navigation
- Refactor refreshDevices() into smaller helper methods
- AudioPriorityBarApp.swift reduced from 441 to 86 lines

File structure now:
- ViewModels/AudioManager.swift (state management)
- Models/StoredDevice.swift (persisted device data)
- Services/PriorityManager.swift (priority/persistence logic)
